+++ docs/releases.rst
@@ -1,5 +1,14 @@
 Releases
 ========
+Version 0.5.0
+-------------
+**Changed**
+
+* Rename ``get_async`` attribute to ``get_page_async``
+  in ``CountOffsetPaginator``;
+* Rename ``get_sync`` attribute to ``get_page_sync``
+  in ``CountOffsetPaginator``;
+
 Version 0.4.0
 -------------
 **Added**
@@ -11,7 +20,7 @@ Version 0.3.0
 -------------
 **Changed**
 
-* Rename ``definition`` attribute to ``discriminator`` in ``PolymorphicMixin``;
+* Rename ``definition`` attribute to ``discriminator`` in ``PolymorphicMixin``.
 
 Version 0.2.0
 -------------